quell warnings (part 3)
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21130 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/AutoPilot.cxx
@@ -469,7 +469,7 @@ static bool chasePlayer(float &rotation, float &speed)
       rotation *= (float)M_PI / (2.0f * fabs(rotation));
       speed = dotProd; //go forward inverse rel to how much you need to turn
     } else {
-      const int64_t period = BzTime::getTick().getSeconds();
+      const int64_t period = (int64_t)BzTime::getTick().getSeconds();
       float absBias = (float)(M_PI/20.0 * (distance / 100.0));
       float bias = ((period % 4) < 2) ? absBias : -absBias;
       rotation += bias;
